subscriber/requirements.txt
eclipse-zenoh==1.0.0
protobuf==5.28.2